PureGo Switch: replace all *C.char instances with string
Purego handles passing strings as long as C doesn't hold on to the
reference, which is the same as in `cgo rules`.

It is also how it has been done in this file previously as well, most
strings are copied into the C memory through CString wrapper and then
freed later, which is exactly how purego handles strings when they are
not null terminated, thus all in this case.

This should not break anything in theory but more tests should be done.