PureGo Switch: remove 'cgo'
Completely remove `cgo`. Remove import statements.

This commit introduces the management of library files, because
previously `cgo` dynamically lined these files to the `file` but purego
links to the process, meaning that they have to be open while the
program is working.

This commit will make the program the load and link the files in the
init and close them when exiting.